\name{add_video_to_playlist}
\alias{add_video_to_playlist}
\title{Add Video to Playlist}
\description{Add Video to Playlist}
\usage{
add_video_to_playlist(playlist_id, video_id, position, ...)
}
\arguments{
  \item{playlist_id}{string; Required. The ID of the playlist.}
  \item{video_id}{string; Required. The ID of the video to add.}
  \item{position}{numeric; Optional. The position of the video in the playlist. If not provided, the video will be added to the end of the playlist.}
  \item{...}{Additional arguments passed to \\code\{\\link\{tuber_POST_json\}\}.}
}
\value{
Details of the added video in the playlist.
}
\references{
\\url\{https://developers.google.com/youtube/v3/docs/playlistItems/insert\}
}
\examples{
\dontrun{

# Set API token via yt_oauth() first

add_video_to_playlist(playlist_id = "YourPlaylistID", video_id = "2_gLD1jarfU")
}
}
