% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_caption.R
\name{delete_caption}
\alias{delete_caption}
\title{Delete a Caption Track}
\usage{
delete_caption(caption_id, ...)
}
\arguments{
\item{caption_id}{Caption-track ID.}

\item{...}{Additional arguments passed to [tuber_DELETE()].}
}
\value{
The empty API response, invisibly.
}
\description{
Delete a Caption Track
}
\examples{
\dontrun{
delete_caption("y3ElXcEME3lSISz6izkWVT5GvxjPu8pA")
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/captions/delete}
}
