% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_channel_section.R
\name{delete_channel_section}
\alias{delete_channel_section}
\title{Delete a Channel Section}
\usage{
delete_channel_section(section_id, ...)
}
\arguments{
\item{section_id}{Channel-section ID.}

\item{...}{Additional arguments passed to [tuber_DELETE()].}
}
\value{
The empty API response, invisibly.
}
\description{
Delete a Channel Section
}
\examples{
\dontrun{
delete_channel_section("SECTION_ID")
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/channelSections/delete}
}
