% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_comment.R
\name{delete_comment}
\alias{delete_comment}
\title{Delete a Comment}
\usage{
delete_comment(comment_id, ...)
}
\arguments{
\item{comment_id}{Comment ID.}

\item{...}{Additional arguments passed to [tuber_DELETE()].}
}
\value{
The empty API response, invisibly.
}
\description{
Delete a Comment
}
\examples{
\dontrun{
delete_comment("COMMENT_ID")
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/comments/delete}
}
