% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_playlist.R
\name{delete_playlist}
\alias{delete_playlist}
\title{Delete a Playlist}
\usage{
delete_playlist(playlist_id, ...)
}
\arguments{
\item{playlist_id}{Playlist ID.}

\item{...}{Additional arguments passed to [tuber_DELETE()].}
}
\value{
The empty API response, invisibly.
}
\description{
Delete a Playlist
}
\examples{
\dontrun{
delete_playlist("PLAYLIST_ID")
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/playlists/delete}
}
