% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_playlist_item.R
\name{delete_playlist_item}
\alias{delete_playlist_item}
\title{Delete a Playlist Item}
\usage{
delete_playlist_item(playlist_item_id, ...)
}
\arguments{
\item{playlist_item_id}{Playlist-item ID, not the video ID.}

\item{...}{Additional arguments passed to [tuber_DELETE()].}
}
\value{
The empty API response, invisibly.
}
\description{
Delete a Playlist Item
}
\examples{
\dontrun{
delete_playlist_item("PLAYLIST_ITEM_ID")
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/playlistItems/delete}
}
