% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_video.R
\name{delete_video}
\alias{delete_video}
\title{Delete a Video}
\usage{
delete_video(video_id, ...)
}
\arguments{
\item{video_id}{Video ID.}

\item{...}{Additional arguments passed to [tuber_DELETE()].}
}
\value{
The empty API response, invisibly.
}
\description{
Delete a Video
}
\examples{
\dontrun{
delete_video("VIDEO_ID")
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/videos/delete}
}
