\name{generate_cache_key}
\alias{generate_cache_key}
\title{Generate cache key for API request}
\description{Generate cache key for API request}
\usage{
generate_cache_key(endpoint, query, auth)
}
\arguments{
  \item{endpoint}{API endpoint name}
  \item{query}{Query parameters}
  \item{auth}{Authentication method}
}
\value{
Character cache key
}
\keyword{internal}
