% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_activities.R
\name{get_activities}
\alias{get_activities}
\title{Get Channel Activity}
\usage{
get_activities(video_id = NULL, lang = "en", ...)
}
\arguments{
\item{video_id}{id of the video; required; no default}

\item{lang}{language of the caption; required; default is english ("en")}

\item{\dots}{Additional arguments passed to \code{\link{tuber_GET}}.}
}
\value{
captions for the video from one of the first track
}
\description{
Get Channel Activity
}
\examples{
\dontrun{
get_activities(video_id="yJXTXN4xrI8")
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/captions/list}
}

