% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_channel.R
\name{get_channel}
\alias{get_channel}
\title{Get statistics of a Channel}
\usage{
get_channel(channel_id = NULL)
}
\arguments{
\item{channel_id}{Character. Id of the channel}
}
\value{
list with 5 elements: viewCount, likeCount, dislikeCount, favoriteCount, commentCount
}
\description{
Get statistics of a Channel
}
\examples{
\dontrun{
get_channel(channel_id="UChTJTbr5kf3hYazJZO-euHg")
}
}
\references{
\url{https://console.developers.google.com/project}
}

