% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_details.R
\name{get_details}
\alias{get_details}
\title{Get Details about a Video}
\usage{
get_details(video_id = NULL)
}
\arguments{
\item{video_id}{id of the video; required}
}
\value{
list with the following elements - publishedAt, channelId, title, description, thumbnails, 
channelTitle, categoryId, liveBroadcastContent, localized, defaultAudioLanguage
}
\description{
Get details such as when the video was published, the title, description, thumbnails, category etc.
}
\examples{
\dontrun{
get_details(video_id="N708P-A45D0")
}
}
\references{
\url{https://developers.google.com/youtube/v3/}
}

