% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_most_comments.R
\name{get_most_comments}
\alias{get_most_comments}
\title{Get Most Comments}
\usage{
get_most_comments(video_id)
}
\arguments{
\item{video_id}{ID of video, required.}
}
\value{
Data frame with all comments and replies.
}
\description{
Retrieves all top level comments and replies to them.
Replies to replies are not included.
}
\examples{
\dontrun{

# Set API token via yt_oauth() first

get_most_comments(video_id = "Hop_MfkXl7c")
}



}
