% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_playlist_item_ids.R
\name{get_playlist_item_ids}
\alias{get_playlist_item_ids}
\title{Get Playlist-Item IDs}
\usage{
get_playlist_item_ids(
  playlist_id = NULL,
  playlist_item_ids = NULL,
  video_id = NULL,
  max_results = 50,
  page_token = NULL,
  auth = "key",
  ...
)
}
\arguments{
\item{playlist_id}{Optional playlist ID.}

\item{playlist_item_ids}{Optional character vector of playlist-item IDs.
Supply exactly one of `playlist_id` or `playlist_item_ids`.}

\item{video_id}{Optional video ID used to filter items in `playlist_id`.}

\item{max_results}{Maximum total number of items to return.}

\item{page_token}{Optional page token at which to start.}

\item{auth}{Authentication method, `"key"` or `"token"`.}

\item{...}{Additional arguments passed to [tuber_GET()].}
}
\value{
A character vector of playlist-item IDs.
}
\description{
Get Playlist-Item IDs
}
\examples{
\dontrun{
get_playlist_item_ids(playlist_id = "PLrEnWoR732-CN09YykVof2lxdI3MLOZda")
}
}
