% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_playlist_video_ids.R
\name{get_playlist_video_ids}
\alias{get_playlist_video_ids}
\title{Get Video IDs from a Playlist}
\usage{
get_playlist_video_ids(
  playlist_id,
  max_results = 50,
  page_token = NULL,
  auth = "key",
  ...
)
}
\arguments{
\item{playlist_id}{Playlist ID.}

\item{max_results}{Maximum total number of video IDs to return.}

\item{page_token}{Optional page token at which to start.}

\item{auth}{Authentication method, `"key"` or `"token"`.}

\item{...}{Additional arguments passed to [list_playlist_items()].}
}
\value{
A character vector of video IDs.
}
\description{
Get Video IDs from a Playlist
}
\examples{
\dontrun{
get_playlist_video_ids("PLrEnWoR732-CN09YykVof2lxdI3MLOZda")
}
}
