% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_playlists.R
\name{get_playlists}
\alias{get_playlists}
\title{Get Playlists}
\usage{
get_playlists(filter = NULL, part = "contentDetails", max_results = 50,
  hl = NULL, page_token = NULL, ...)
}
\arguments{
\item{filter}{string; Required.
named vector of length 1
potential names of the entry in the vector: 
\code{channel_id}: ID of the channel
\code{playlist_id}: YouTube playlist ID.}

\item{part}{Required. One of the following: \code{contentDetails, id, localizations, player, snippet, status}. Default: \code{contentDetails}.}

\item{max_results}{Maximum number of items that should be returned. Integer. Optional. Can be between 0 and 50. Default is 50.}

\item{hl}{Language used for text values. Optional. Default is \code{en-US}. For other allowed language codes, see \code{\link{list_langs}}.}

\item{page_token}{specific page in the result set that should be returned, optional}

\item{\dots}{Additional arguments passed to \code{\link{tuber_GET}}.}
}
\value{
captions for the video from one of the first track
}
\description{
Get Playlists
}
\examples{
\dontrun{

# Set API token via yt_oauth() first

get_playlists(filter=c(channel_id="UChTJTbr5kf3hYazJZO-euHg"))
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/playlists/list}
}

