% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_items}
\alias{has_items}
\title{Check if object has items}
\usage{
has_items(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical indicating if x has items
}
\description{
Returns TRUE if x is not NULL and has length > 0
}
\keyword{internal}
