% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{is_cacheable_endpoint}
\alias{is_cacheable_endpoint}
\title{Check if endpoint should be cached}
\usage{
is_cacheable_endpoint(endpoint)
}
\arguments{
\item{endpoint}{API endpoint name}
}
\value{
Logical indicating if endpoint is cacheable
}
\description{
Check if endpoint should be cached
}
\keyword{internal}
