\name{is_transient_error}
\alias{is_transient_error}
\title{Check if an error is transient and worth retrying}
\description{Check if an error is transient and worth retrying}
\usage{
is_transient_error(error)
}
\arguments{
  \item{error}{Error object to check}
}
\value{
Logical indicating if error is transient
}
\keyword{internal}
