% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_channel_resources.R
\name{list_channel_resources}
\alias{list_channel_resources}
\title{Returns List of Requested Channel Resources}
\usage{
list_channel_resources(channel_name = NULL, part = "contentDetails", ...)
}
\arguments{
\item{channel_name}{string, channel name, required}

\item{part}{a comma separated list of channel resource properties that response will include
string, required, one of the following: auditDetails, brandingSettings, contentDetails, contentOwnerDetails, id, invideoPromotion, localizations, snippet, statistics, status, topicDetails
default is contentDetails}

\item{\dots}{Additional arguments passed to \code{\link{tuber_GET}}.}
}
\value{
list
}
\description{
Returns List of Requested Channel Resources
}
\examples{
\dontrun{
list_channel_videos("latenight")
list_channel_videos("latenight", part="id, contentDetails")
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/channels/list}
}

