% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_langs.R
\name{list_langs}
\alias{list_langs}
\title{List Languages That Youtube Currently Supports}
\usage{
list_langs()
}
\value{
data.frame with 3 columns: hl (two letter abbreviation), name (of the language), etag
}
\description{
List Languages That Youtube Currently Supports
}
\examples{
\dontrun{
list_langs()
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/i18nLanguages/list}
}

