\name{list_langs_cached}
\alias{list_langs_cached}
\title{List supported languages with caching}
\description{List supported languages with caching}
\usage{
list_langs_cached(auth, cache_ttl, ...)
}
\arguments{
  \item{auth}{Authentication method}
  \item{cache_ttl}{Cache time-to-live (default: 24 hours)}
  \item{...}{Additional arguments}
}
\value{
Languages data
}
