\name{list_regions_cached}
\alias{list_regions_cached}
\title{List supported regions with caching}
\description{List supported regions with caching}
\usage{
list_regions_cached(auth, cache_ttl, ...)
}
\arguments{
  \item{auth}{Authentication method}
  \item{cache_ttl}{Cache time-to-live (default: 24 hours)}
  \item{...}{Additional arguments}
}
\value{
Regions data
}
