% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quota_management.R
\name{quota_cost}
\alias{quota_cost}
\title{Track Quota Usage}
\usage{
quota_cost(endpoint, method)
}
\arguments{
\item{endpoint}{Character. API resource name.}

\item{method}{Character. API method such as `"list"`, `"insert"`,
`"update"`, `"delete"`, or `"download"`.}
}
\description{
Internal function to track API usage
}
\keyword{internal}
