% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_extract}
\alias{safe_extract}
\title{Safely extract field from list/object}
\usage{
safe_extract(obj, field, default = NA_character_)
}
\arguments{
\item{obj}{Object to extract from}

\item{field}{Field name (character)}

\item{default}{Default value if field missing. Default: NA_character_}
}
\value{
The field value or default
}
\description{
Extracts a field from an object, returning a default value if the field
is missing or NULL.
}
\keyword{internal}
