% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_nested}
\alias{safe_nested}
\title{Safely extract nested field}
\usage{
safe_nested(obj, ..., default = NA_character_)
}
\arguments{
\item{obj}{Object to extract from}

\item{...}{Field names in order (e.g., "author", "id", "value")}

\item{default}{Default value if any field missing. Default: NA_character_}
}
\value{
The nested field value or default
}
\description{
Extracts a value from nested list structures, returning a default value
if any level of the path is missing.
}
\keyword{internal}
