% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{store_cached_response}
\alias{store_cached_response}
\title{Store response in cache}
\usage{
store_cached_response(cache_key, data, ttl = NULL)
}
\arguments{
\item{cache_key}{Cache key}

\item{data}{Response data to cache}

\item{ttl}{Time-to-live in seconds (NULL for default)}
}
\description{
Store response in cache
}
\keyword{internal}
