% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{suggest_solution}
\alias{suggest_solution}
\title{Provide helpful suggestions for common user errors}
\usage{
suggest_solution(issue_type, details = "")
}
\arguments{
\item{issue_type}{Type of issue encountered}

\item{details}{Additional details for the suggestion}
}
\description{
Provide helpful suggestions for common user errors
}
\keyword{internal}
