% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuber.R
\docType{package}
\name{tuber-package}
\alias{tuber-package}
\title{tuber: Client for the YouTube Data API}
\description{
Search public YouTube data and retrieve channels, videos, playlists, comments, captions, live broadcasts, and reference data. Authenticated methods support common uploads, playlist changes, comment moderation, and media updates. See the YouTube Data API documentation at \url{https://developers.google.com/youtube/v3/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gojiplus.github.io/tuber/}
  \item \url{https://github.com/gojiplus/tuber}
  \item Report bugs at \url{https://github.com/gojiplus/tuber/issues}
}

}
\author{
\strong{Maintainer}: Gaurav Sood \email{gsood07@gmail.com}

Authors:
\itemize{
  \item Gaurav Sood \email{gsood07@gmail.com}
}

Other contributors:
\itemize{
  \item Kate Lyons \email{k.lyons7@gmail.com} [contributor]
  \item John Muschelli \email{muschellij2@gmail.com} [contributor]
}

}
\keyword{internal}
