\name{tuber_cache_clear}
\alias{tuber_cache_clear}
\title{Clear cache entries}
\description{Clear cache entries}
\usage{
tuber_cache_clear(pattern, older_than)
}
\arguments{
  \item{pattern}{Regular expression pattern to match cache keys (NULL for all)}
  \item{older_than}{Clear entries older than this many seconds}
}
