% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuber.R
\name{tuber_perform}
\alias{tuber_perform}
\title{Perform a request and check the response}
\usage{
tuber_perform(req)
}
\arguments{
\item{req}{An httr2 request}
}
\value{
An httr2 response
}
\description{
Perform a request and check the response
}
\keyword{internal}
