% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuber.R
\name{tuber_POST_json}
\alias{tuber_POST_json}
\title{POST encoded in json}
\usage{
tuber_POST_json(path, query, body = "", ...)
}
\arguments{
\item{path}{path to specific API request URL}

\item{query}{query list}

\item{body}{passing image through body}

\item{\dots}{Additional arguments passed to \code{\link[httr]{GET}}.}
}
\value{
list
}
\description{
POST encoded in json
}
\keyword{internal}
