% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuber.R
\name{tuber_PUT}
\alias{tuber_PUT}
\title{PUT}
\usage{
tuber_PUT(path, query, body = "", ...)
}
\arguments{
\item{path}{path to specific API request URL}

\item{query}{query list}

\item{body}{list serialized as the JSON request body}

\item{\dots}{Ignored; see [tuber_POST_json()].}
}
\value{
list
}
\description{
PUT
}
\keyword{internal}
