% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuber.R
\name{tuber_upload_body}
\alias{tuber_upload_body}
\title{Send the file to a resumable upload session}
\usage{
tuber_upload_body(upload_url, file, type)
}
\arguments{
\item{upload_url}{Session URL from [tuber_upload_session()]}

\item{file}{path to the file to upload}

\item{type}{MIME type of \code{file}}
}
\value{
An httr2 response
}
\description{
The second half of [tuber_upload_session()]. The session URL is a full URL
YouTube chose, so this cannot go through [tuber_request()].
}
\keyword{internal}
