% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuber.R
\name{tuber_upload_session}
\alias{tuber_upload_session}
\title{Open a resumable media upload session}
\usage{
tuber_upload_session(path, query, metadata, file, type)
}
\arguments{
\item{path}{API endpoint path under \code{upload/youtube/v3}}

\item{query}{query list}

\item{metadata}{list serialized as the JSON metadata body}

\item{file}{path to the file that will be uploaded}

\item{type}{MIME type of \code{file}}
}
\value{
The session URL as a string
}
\description{
YouTube's resumable protocol is two requests: a metadata POST that returns a
session URL in the Location header, then a PUT of the bytes to that URL.
This is the first half.
}
\keyword{internal}
