% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuber.R
\name{tuber_write}
\alias{tuber_write}
\title{Write request with a JSON body}
\usage{
tuber_write(path, method, query, body = NULL, quota_method)
}
\arguments{
\item{path}{path to specific API request URL}

\item{method}{HTTP method, one of "POST", "PUT" or "DELETE"}

\item{query}{query list}

\item{body}{list serialized as the JSON request body}

\item{quota_method}{quota category recorded for this call}
}
\value{
An httr2 response
}
\description{
Write request with a JSON body
}
\keyword{internal}
