% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_related_videos.R
\name{yt_get_related_videos}
\alias{yt_get_related_videos}
\title{Get Related Videos}
\usage{
yt_get_related_videos(video_id = NULL, maxResults = 5,
  safeSearch = "none", ...)
}
\arguments{
\item{video_id}{Character. Required. No default.}

\item{maxResults}{Numeric. 0 to 50. Acceptable values are 0 to 50, inclusive.}

\item{safeSearch}{Character. Optional. Takes one of three values: 'moderate', 'none' (default) or 'strict'}

\item{\dots}{Additional arguments passed to \code{\link{tuber_GET}}.}
}
\value{
data.frame with 7 columns: publishedAt, channelId, title, description, thumbnails, channelTitle, liveBroadcastContent
}
\description{
Takes a video id and returns related videos
}
\examples{
\dontrun{
yt_get_related_videos(video_id="yJXTXN4xrI8")
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/search/list}
}

