% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_list_guidecats.R
\name{yt_list_guidecats}
\alias{yt_list_guidecats}
\title{Get List of categories that can be associated with YouTube channels}
\usage{
yt_list_guidecats(regionCode = NULL)
}
\arguments{
\item{regionCode}{Character. Required. Has to be a ISO 3166-1 alpha-2 code (see \url{https://www.iso.org/obp/ui/#search}).}
}
\value{
data.frame with 4 columns: channelId, title, etag, id
}
\description{
Get List of categories that can be associated with YouTube channels
}
\examples{
\dontrun{
yt_list_guidecats("JP")
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/search/list}
}

