% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_list_langs.R
\name{yt_list_langs}
\alias{yt_list_langs}
\title{List Languages That Youtube Currently Supports}
\usage{
yt_list_langs()
}
\value{
data.frame with 2 columns: hl (two letter abbreviation), name (of the language)
}
\description{
List Languages That Youtube Currently Supports
}
\examples{
\dontrun{
yt_list_langs()
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/i18nLanguages/list}
}

