% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_list_regions.R
\name{yt_list_regions}
\alias{yt_list_regions}
\title{List Content Regions That Youtube Currently Supports}
\usage{
yt_list_regions()
}
\value{
data.frame with 2 columns: gl (two letter abbreviation), name (of the region)
}
\description{
List Content Regions That Youtube Currently Supports
}
\examples{
\dontrun{
yt_list_regions()
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/i18nRegions/list}
}

