% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_list_videocats.R
\name{yt_list_videocats}
\alias{yt_list_videocats}
\title{List of Categories That Can be Associated with Videos}
\usage{
yt_list_videocats(regionCode = NULL)
}
\arguments{
\item{regionCode}{Character. Required. Has to be a ISO 3166-1 alpha-2 code (see \url{https://www.iso.org/obp/ui/#search}).}
}
\value{
data.frame with 3 columns - channelId, title, assignable
}
\description{
List of Categories That Can be Associated with Videos
}
\examples{
\dontrun{
yt_list_videocats("JP")
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/search/list}
}

