% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quota_management.R
\name{yt_reset_quota}
\alias{yt_reset_quota}
\title{Reset Quota Counter}
\usage{
yt_reset_quota()
}
\description{
Reset the quota counter (typically done automatically at midnight Pacific
Time, when YouTube's daily quota resets)
}
