% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_topic_search.R
\name{yt_topic_search}
\alias{yt_topic_search}
\title{Search YouTube by Topic
It uses the Freebase list of topics}
\usage{
yt_topic_search(topic = NULL)
}
\arguments{
\item{topic}{topic being searched for; required; no default}
}
\value{
a list
}
\description{
Search YouTube by Topic
It uses the Freebase list of topics
}
\examples{
 \dontrun{yt_topic_search(topic="Barack Obama")}
}
\references{
\url{http://www.freebase.com/}
}

