% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuber.R
\name{add_tuber_attributes}
\alias{add_tuber_attributes}
\title{Add standardized metadata attributes to API response}
\usage{
add_tuber_attributes(
  result,
  api_calls_made = 1,
  quota_used = NULL,
  function_name = NULL,
  parameters = list(),
  timestamp = Sys.time(),
  ...
)
}
\arguments{
\item{result}{The result object to add attributes to}

\item{api_calls_made}{Number of API calls made to generate this result}

\item{quota_used}{Estimated quota units consumed}

\item{function_name}{Name of the calling function}

\item{parameters}{List of key parameters used in the function call}

\item{timestamp}{When the API call was made}

\item{...}{Additional custom attributes}
}
\value{
The result object with standardized attributes added
}
\description{
Adds consistent metadata attributes to function return values for better
debugging and quota management.
}
\keyword{internal}
