% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_video_to_playlist.R
\name{add_video_to_playlist}
\alias{add_video_to_playlist}
\title{Add Video to Playlist}
\usage{
add_video_to_playlist(playlist_id, video_id, position = NULL, ...)
}
\arguments{
\item{playlist_id}{string; Required. The ID of the playlist.}

\item{video_id}{string; Required. The ID of the video to add.}

\item{...}{Additional arguments passed to \code{\link{tuber_POST_json}}.}
}
\value{
Details of the added video in the playlist.
}
\description{
Add Video to Playlist
}
\examples{
\dontrun{

# Set API token via yt_oauth() first

add_video_to_playlist(playlist_id = "YourPlaylistID", video_id = "2_gLD1jarfU")
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/playlistItems/insert}
}
