% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_playlist_title.R
\name{change_playlist_title}
\alias{change_playlist_title}
\title{Change the title of a YouTube playlist.}
\usage{
change_playlist_title(playlist_id, new_title)
}
\arguments{
\item{playlist_id}{A character string specifying the ID of the playlist you want to update.}

\item{new_title}{A character string specifying the new title for the playlist.}
}
\value{
A list containing the server response after the update attempt.
}
\description{
This function updates the title of an existing YouTube playlist using the YouTube Data API.
}
\examples{
\dontrun{
change_playlist_title(playlist_id = "YourPlaylistID", new_title = "New Playlist Title")
}

}
