% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_playlist.R
\name{create_playlist}
\alias{create_playlist}
\title{Create New Playlist}
\usage{
create_playlist(title, description = "", status = "public", ...)
}
\arguments{
\item{title}{string; Required. The title of the playlist.}

\item{description}{string; Optional. The description of the playlist.}

\item{status}{string; Optional. Default: 'public'. Can be one of: 'private', 'public', or
'unlisted'.}

\item{...}{Ignored; retained for backward compatibility.}
}
\value{
The created playlist's details.
}
\description{
Create New Playlist
}
\examples{
\dontrun{

# Set API token via yt_oauth() first

create_playlist(title = "My New Playlist", description = "This is a test playlist.")
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/playlists/insert}
}
