% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_captions.R
\name{delete_captions}
\alias{delete_captions}
\title{Delete a Particular Caption Track}
\usage{
delete_captions(id = NULL, ...)
}
\arguments{
\item{id}{String. Required. id of the caption track that is being retrieved}

\item{\dots}{Additional arguments passed to \code{\link{tuber_DELETE}}.}
}
\description{
Delete a Particular Caption Track
}
\examples{
\dontrun{

# Set API token via yt_oauth() first

delete_captions(id = "y3ElXcEME3lSISz6izkWVT5GvxjPu8pA")
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/captions/delete}
}
