% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_channel_sections.R
\name{delete_channel_sections}
\alias{delete_channel_sections}
\title{Delete Channel Sections}
\usage{
delete_channel_sections(id = NULL, ...)
}
\arguments{
\item{id}{Required. ID of the channel section.}

\item{\dots}{Additional arguments passed to \code{\link{tuber_DELETE}}.}
}
\value{
captions for the video from one of the first track
}
\description{
Delete a Channel Section
}
\examples{

\dontrun{

# Set API token via yt_oauth() first

delete_channel_sections(c(channel_id = "UCRw8bIz2wMLmfgAgWm903cA"))
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/channelSections/delete}
}
