% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_playlist_items.R
\name{delete_playlist_items}
\alias{delete_playlist_items}
\title{Delete a Playlist Item}
\usage{
delete_playlist_items(id = NULL, ...)
}
\arguments{
\item{id}{String. Required. id of the playlist item that is to be deleted}

\item{\dots}{Additional arguments passed to \code{\link{tuber_DELETE}}.}
}
\description{
Delete a Playlist Item
}
\examples{
\dontrun{

# Set API token via yt_oauth() first

delete_playlist_items(id = "YourPlaylistItemID")
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/playlistItems/delete}
}
