% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_videos.R
\name{delete_videos}
\alias{delete_videos}
\title{Delete a Video}
\usage{
delete_videos(id = NULL, ...)
}
\arguments{
\item{id}{String. Required. id of the video that is to be deleted}

\item{\dots}{Additional arguments passed to \code{\link{tuber_DELETE}}.}
}
\value{

}
\description{
Delete a Video
}
\examples{
\dontrun{

# Set API token via yt_oauth() first

delete_videos(id = "y3ElXcEME3lSISz6izkWVT5GvxjPu8pA")
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/playlistItems/delete}
}
