% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{generate_cache_key}
\alias{generate_cache_key}
\title{Generate cache key for API call}
\usage{
generate_cache_key(endpoint, query, auth)
}
\arguments{
\item{endpoint}{API endpoint name}

\item{query}{Query parameters}

\item{auth}{Authentication method}
}
\value{
Character cache key
}
\description{
Generate cache key for API call
}
