\name{get_all_channel_video_stats}
\alias{get_all_channel_video_stats}
\title{Get statistics on all the videos in a Channel}
\description{Efficiently collects all video IDs from a channel's uploads playlist, then fetches statistics and details using batch processing for optimal API quota usage.}
\usage{
get_all_channel_video_stats(channel_id, mine, ...)
}
\arguments{
  \item{channel_id}{Character. Id of the channel}
  \item{mine}{Boolean. TRUE if you want to fetch stats of your own channel. Default is FALSE.}
  \item{\\dots}{Additional arguments passed to \\code\{\\link\{tuber_GET\}\}.}
}
\value{
A \\code\{data.frame\} containing video metadata along with view, like,
  dislike and comment counts.

If the \\code\{channel_id\} is mistyped or there is no information, an empty list is returned
}
\references{
\\url\{https://developers.google.com/youtube/v3/docs/channels/list\}
}
\examples{
\dontrun{

# Set API token via yt_oauth() first

get_all_channel_video_stats(channel_id="UCxOhDvtaoXDAB336AolWs3A")
get_all_channel_video_stats(channel_id="UCMtFAi84ehTSYSE9Xo") # Incorrect channel ID
}
}
