\name{get_all_comments}
\alias{get_all_comments}
\title{Get all the comments for a video including replies}
\description{Get all the comments for a video including replies}
\usage{
get_all_comments(video_id, ...)
}
\arguments{
  \item{video_id}{string; Required. \\code\{video_id\}: video ID.}
  \item{\\dots}{Additional arguments passed to \\code\{\\link\{tuber_GET\}\}.}
}
\value{
a \\code\{data.frame\} with the following columns:
\\code\{authorDisplayName, authorProfileImageUrl, authorChannelUrl,\}
\\code\{ authorChannelId.value, videoId, textDisplay,
canRate, viewerRating, likeCount, publishedAt, updatedAt,
id, moderationStatus, parentId\}
}
\references{
\\url\{https://developers.google.com/youtube/v3/docs/commentThreads/list\}
}
\examples{
\dontrun{

# Set API token via yt_oauth() first

get_all_comments(video_id = "a-UQz7fqR3w")
}
}
