% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_captions.R
\name{get_captions}
\alias{get_captions}
\title{Get Captions of a Video}
\usage{
get_captions(video_id = NULL, part = "snippet", id = NULL, ...)
}
\arguments{
\item{video_id}{ID of the video whose captions are requested. Required. No default.}

\item{part}{Required. Default is \code{snippet}. It can also be \code{id}.}

\item{id}{String. Optional. id of the caption track that is being retrieved}

\item{\dots}{Additional arguments passed to \code{\link{tuber_GET}}.}
}
\value{
String.
}
\description{
For getting captions from the v3 API, you must specify the id resource. 
Check \code{\link{list_caption_tracks}} for more information.
}
\examples{
\dontrun{

# Set API token via yt_oauth() first

get_captions(video_id = "yJXTXN4xrI8")
}
}
