% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_captions.R
\name{get_captions}
\alias{get_captions}
\title{Get Captions of a Video}
\usage{
get_captions(part = "snippet", video_id = NULL, lang = "en", id = NULL,
  ...)
}
\arguments{
\item{part}{Caption resource requested. Required. Comma separated list of one or more of the 
following: id, snippet. e.g., "id, snippet", "id" Default: snippet.}

\item{video_id}{ID of the video whose captions are requested. Required. No default.}

\item{lang}{Language of the caption; required; default is english ("en")}

\item{id}{comma-separated list of IDs that identify the caption resources that should be retrieved; optional; string}

\item{\dots}{Additional arguments passed to \code{\link{tuber_GET}}.}
}
\value{
captions for the video from one of the first track
}
\description{
Get Captions of a Video
}
\examples{
\dontrun{
get_captions(video_id="yJXTXN4xrI8")
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/captions/list}
}

