% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_captions.R
\name{get_captions}
\alias{get_captions}
\title{Get Captions of a Video}
\usage{
get_captions(video_id = NULL, lang = "en")
}
\arguments{
\item{video_id}{id of the video; required; no default}

\item{lang}{language of the caption; required; default is english ("en")}
}
\value{
list with 5 elements - viewCount, likeCount, dislikeCount, favoriteCount, commentCount
}
\description{
Get Captions of a Video
}
\examples{
 \dontrun{get_captions(video_id="yJXTXN4xrI8")}
}
\references{
\url{https://console.developers.google.com/project}
}

