% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{get_channel_info_cached}
\alias{get_channel_info_cached}
\title{Get channel information with caching (for static parts)}
\usage{
get_channel_info_cached(
  channel_id,
  part = "snippet,brandingSettings",
  auth = "key",
  cache_ttl = 3600,
  ...
)
}
\arguments{
\item{channel_id}{Channel ID}

\item{part}{Parts to retrieve (only static parts will be cached)}

\item{auth}{Authentication method}

\item{cache_ttl}{Cache time-to-live (default: 1 hour for channel info)}

\item{...}{Additional arguments}
}
\value{
Channel information
}
\description{
Get channel information with caching (for static parts)
}
