% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_channel_stats.R
\name{get_channel_stats}
\alias{get_channel_stats}
\title{Get statistics of a Channel}
\usage{
get_channel_stats(channel_id = NULL, ...)
}
\arguments{
\item{channel_id}{Character. Id of the channel}

\item{\dots}{Additional arguments passed to \code{\link{tuber_GET}}.}
}
\value{
list with 5 elements: viewCount, likeCount, dislikeCount, favoriteCount, commentCount
}
\description{
Get statistics of a Channel
}
\examples{
\dontrun{
get_channel_stats(channel_id="UChTJTbr5kf3hYazJZO-euHg")
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/channels/list}
}

