% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comments.R
\name{get_comments}
\alias{get_comments}
\title{Get Comments On a Video}
\usage{
get_comments(video_id = NULL)
}
\arguments{
\item{video_id}{id of the video; required}
}
\value{
Nested named list. The entry \code{items} is a list of comments along with meta information. 
Within each of the \code{items} is an item \code{snippet} which has an item \code{topLevelComment$snippet$textDisplay}
that contains the actual comment.
}
\description{
Get Comments On a Video
}
\examples{
\dontrun{
get_comments(video_id="N708P-A45D0")
}
}
\references{
\url{https://console.developers.google.com/project}
}

