% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_comments.R
\name{get_comments}
\alias{get_comments}
\title{Get Statistics of a Video}
\usage{
get_comments(video_id = NULL)
}
\arguments{
\item{video_id}{id of the video; required}
}
\value{
list with 5 elements - viewCount, likeCount, dislikeCount, favoriteCount, commentCount
}
\description{
Get Statistics of a Video
}
\examples{
get_comments(video_id="N708P-A45D0")
}
\references{
\url{https://console.developers.google.com/project}
}

