% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comments.R
\name{get_comments}
\alias{get_comments}
\title{Get Comments To a Video}
\usage{
get_comments(video_id = NULL)
}
\arguments{
\item{video_id}{id of the video; required}
}
\value{
XML list of comments
}
\description{
Get Comments To a Video
}
\examples{
 \dontrun{get_comments(video_id="N708P-A45D0")}
}
\references{
\url{https://console.developers.google.com/project}
}

