% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_playlist_items.R
\name{get_playlist_items}
\alias{get_playlist_items}
\title{Get Playlist Items}
\usage{
get_playlist_items(playlist_id = NULL, part = "contentDetails",
  max_results = 50, video_id = NULL, page_token = NULL, ...)
}
\arguments{
\item{playlist_id}{ID of the playlist}

\item{part}{Required. Comma separated string including one or more of the following: contentDetails, id, snippet, status. Default: contentDetails.}

\item{max_results}{maximum number of items that should be returned, integer, can be between 0 and 50, default is 50}

\item{video_id}{Optional. request should return only the playlist items that contain the specified video.}

\item{page_token}{specific page in the result set that should be returned, optional}

\item{\dots}{Additional arguments passed to \code{\link{tuber_GET}}.}
}
\value{
captions for the video from one of the first track
}
\description{
Get Playlist Items
}
\examples{
\dontrun{
get_playlist_items(playlist_id="PLrEnWoR732-CN09YykVof2lxdI3MLOZda")
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/playlists/list}
}

