% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats}
\alias{get_stats}
\title{Get statistics of a Video}
\usage{
get_stats(video_id = NULL)
}
\arguments{
\item{video_id}{id of the video; required}
}
\value{
list with 5 elements - viewCount, likeCount, dislikeCount, favoriteCount, commentCount
}
\description{
Get statistics of a Video
}
\examples{
 \dontrun{get_stats(video_id="N708P-A45D0")}
}
\references{
\url{https://console.developers.google.com/project}
}

