% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats}
\alias{get_stats}
\title{Get statistics of a Video}
\usage{
get_stats(video_id = NULL, ...)
}
\arguments{
\item{video_id}{Character. Id of the video. Required.}

\item{\dots}{Additional arguments passed to \code{\link{tuber_GET}}.}
}
\value{
list with 6 elements: \code{id, viewCount, likeCount,
dislikeCount, favoriteCount, commentCount}
}
\description{
Get statistics of a Video
}
\examples{
\dontrun{

# Set API token via yt_oauth() first

get_stats(video_id="N708P-A45D0")
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/videos/list#parameters}
}
