\name{is_static_query}
\alias{is_static_query}
\title{Check if query parameters indicate static data}
\description{Check if query parameters indicate static data}
\usage{
is_static_query(endpoint, query)
}
\arguments{
  \item{endpoint}{API endpoint}
  \item{query}{Query parameters}
}
\value{
Logical indicating if this specific query is cacheable
}
