% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_abuse_report_reasons.R
\name{list_abuse_report_reasons}
\alias{list_abuse_report_reasons}
\title{List reasons that can be used to report abusive videos}
\usage{
list_abuse_report_reasons(...)
}
\arguments{
\item{\dots}{Additional arguments passed to \code{\link{tuber_GET}}.}
}
\value{
data.frame with 3 columns: etag, label, secReasons
}
\description{
List reasons that can be used to report abusive videos
}
\examples{
\dontrun{
list_abuse_report_reasons()
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/videoAbuseReportReasons/list}
}

