% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_captions.R
\name{list_captions}
\alias{list_captions}
\title{List Captions for YouTube Video}
\usage{
list_captions(video_id, query = NULL, auth = "token", ...)
}
\arguments{
\item{video_id}{ID of the YouTube video}

\item{query}{Fields for `query` in `GET`}

\item{auth}{Authentication method: "token" (OAuth2) or "key" (API key)}

\item{...}{Additional arguments to send to \code{\link{tuber_GET}}}
}
\value{
A list containing caption information
}
\description{
List Captions for YouTube Video
}
\examples{
\dontrun{
video_id <- "M7FIvfx5J10"
list_captions(video_id)
}
}
