% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_captions.R
\name{list_captions}
\alias{list_captions}
\title{Upload Video to Youtube}
\usage{
list_captions(video_id, query = NULL, ...)
}
\arguments{
\item{video_id}{ID of the YouTube video}

\item{query}{Fields for `query` in `GET`}

\item{...}{Additional arguments to send to \code{\link{tuber_GET}} and
therefore \code{\link[httr]{GET}}}
}
\value{
A list of the response object from the \code{GET} and content
about captions
}
\description{
Upload Video to Youtube
}
\examples{
\dontrun{
video_id <- "M7FIvfx5J10"
list_captions(video_id)
}
}
