% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_channel_activities.R
\name{list_channel_activities}
\alias{list_channel_activities}
\title{List Channel Activity}
\usage{
list_channel_activities(channel_id = NULL, part = "snippet", ...)
}
\arguments{
\item{channel_id}{id of the channel; required; no default}

\item{part}{specify which part do you want. It can only be one of the three: contentDetails, id, snippet. Default is snippet.}

\item{\dots}{Additional arguments passed to \code{\link{tuber_GET}}.}
}
\value{
captions for the video from one of the first track
}
\description{
Returns a list of channel activity events that match the request criteria.
}
\examples{
\dontrun{
list_channel_activities(channel_id="UCRw8bIz2wMLmfgAgWm903cA")
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/activities/list}
}

