% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_channel_resources.R
\name{list_channel_resources}
\alias{list_channel_resources}
\title{Returns List of Requested Channel Resources}
\usage{
list_channel_resources(channel_name = NULL, part = "contentDetails",
  max_results = 50, page_token = NULL, hl = NULL, ...)
}
\arguments{
\item{channel_name}{string, channel name, required}

\item{part}{a comma separated list of channel resource properties that response will include string. Required. 
One of the following: \code{auditDetails, brandingSettings, contentDetails, contentOwnerDetails, id, invideoPromotion, localizations, snippet, statistics, status, topicDetails}.
Default is \code{contentDetails}.}

\item{max_results}{Maximum number of items that should be returned. Integer. Optional. Can be between 0 and 50. Default is 50.}

\item{page_token}{specific page in the result set that should be returned, optional}

\item{hl}{Language used for text values. Optional. Default is \code{en-US}. For other allowed language codes, see \code{\link{list_langs}}.}

\item{\dots}{Additional arguments passed to \code{\link{tuber_GET}}.}
}
\value{
list
}
\description{
Returns List of Requested Channel Resources
}
\examples{
\dontrun{
list_channel_videos("latenight")
list_channel_videos("latenight", part="id, contentDetails")
list_channel_videos("latenight", part="id, contentDetails", max_results=10)
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/channels/list}
}

