% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_channel_sections.R
\name{list_channel_sections}
\alias{list_channel_sections}
\title{List Channel Sections}
\usage{
list_channel_sections(filter = NULL, part = "snippet", hl = NULL, ...)
}
\arguments{
\item{filter}{string; Required.
named vector of length 1
potential names of the entry in the vector: 
\code{channel_id}: Channel ID
\code{id}: Section ID}

\item{part}{specify which part do you want. It can only be one of the following: \code{contentDetails, id, localizations, snippet, targeting}. Default is \code{snippet}.}

\item{hl}{language that will be used for text values, optional, default is en-US. See also \code{\link{list_langs}}}

\item{\dots}{Additional arguments passed to \code{\link{tuber_GET}}.}
}
\value{
captions for the video from one of the first track
}
\description{
Returns list of channel sections that channel id belongs to.
}
\examples{

\dontrun{

# Set API token via yt_oauth() first

list_channel_sections(c(channel_id="UCRw8bIz2wMLmfgAgWm903cA"))
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/activities/list}
}
