% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_channel_vids.R
\name{list_channel_videos}
\alias{list_channel_videos}
\title{List All the Videos in a Channel}
\usage{
list_channel_videos(channel_name = NULL, ...)
}
\arguments{
\item{channel_name}{string, channel name, required}

\item{\dots}{Additional arguments passed to \code{\link{tuber_GET}}.}
}
\value{
data.frame with 3 columns: hl (two letter abbreviation), name (of the language), etag
}
\description{
List All the Videos in a Channel
}
\examples{
\dontrun{
list_channel_videos("latenight")
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/channels/list}
}

