% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_guidecats.R
\name{list_guidecats}
\alias{list_guidecats}
\title{Get List of categories that can be associated with YouTube channels}
\usage{
list_guidecats(regionCode = NULL, hl = NULL, ...)
}
\arguments{
\item{regionCode}{Character. Required. Has to be a ISO 3166-1 alpha-2 code (see \url{https://www.iso.org/obp/ui/#search})}

\item{hl}{Language used for text values. Optional. Default is \code{en-US}. For other allowed language codes, see \code{\link{list_langs}}.}

\item{\dots}{Additional arguments passed to \code{\link{tuber_GET}}.}
}
\value{
data.frame with 4 columns: channelId, title, etag, id
}
\description{
Get List of categories that can be associated with YouTube channels
}
\examples{
\dontrun{
list_guidecats("JP")
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/search/list}
}

