% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{list_langs_cached}
\alias{list_langs_cached}
\title{List supported languages with caching}
\usage{
list_langs_cached(auth = "key", cache_ttl = 86400, ...)
}
\arguments{
\item{auth}{Authentication method}

\item{cache_ttl}{Cache time-to-live (default: 24 hours)}

\item{...}{Additional arguments}
}
\value{
Languages data
}
\description{
List supported languages with caching
}
