% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_regions.R
\name{list_regions}
\alias{list_regions}
\title{List Content Regions That Youtube Currently Supports}
\usage{
list_regions(hl = NULL, ...)
}
\arguments{
\item{hl}{Language used for text values. Optional. Default is \code{en-US}. For other allowed language codes, see \code{\link{list_langs}}.}

\item{\dots}{Additional arguments passed to \code{\link{tuber_GET}}.}
}
\value{
data.frame with 3 columns: gl (two letter abbreviation), name (of the region), etag
}
\description{
List Content Regions That Youtube Currently Supports
}
\examples{
\dontrun{
list_regions()
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/i18nRegions/list}
}

