% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{list_regions_cached}
\alias{list_regions_cached}
\title{List supported regions with caching}
\usage{
list_regions_cached(auth = "key", cache_ttl = 86400, ...)
}
\arguments{
\item{auth}{Authentication method}

\item{cache_ttl}{Cache time-to-live (default: 24 hours)}

\item{...}{Additional arguments}
}
\value{
Regions data
}
\description{
List supported regions with caching
}
