% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_video_thumbnail.R
\name{set_video_thumbnail}
\alias{set_video_thumbnail}
\title{Set Video Thumbnail}
\usage{
set_video_thumbnail(video_id, file, ...)
}
\arguments{
\item{video_id}{Character. ID of the video to set the thumbnail for.}

\item{file}{Character. Path to the thumbnail image file (JPG or PNG, max 2MB).}

\item{\dots}{Additional arguments passed to \code{\link[httr]{POST}}.}
}
\value{
A list containing the response from the API.
}
\description{
Uploads a custom video thumbnail to YouTube and sets it for a video.
Requires OAuth 2.0 authentication.
}
\examples{
\dontrun{
# Set API token via yt_oauth() first

set_video_thumbnail(video_id = "yJXTXN4xrI8", file = "thumbnail.jpg")
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/thumbnails/set}
}
