\name{store_cached_response}
\alias{store_cached_response}
\title{Store response in cache}
\description{Store response in cache}
\usage{
store_cached_response(cache_key, data, ttl)
}
\arguments{
  \item{cache_key}{Cache key}
  \item{data}{Response data to cache}
  \item{ttl}{Time-to-live in seconds (NULL for default)}
}
