% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuber.R
\name{[.tuber_result}
\alias{[.tuber_result}
\title{Subset method for tuber results}
\usage{
\method{[}{tuber_result}(x, ...)
}
\arguments{
\item{x}{A tuber_result object}

\item{...}{Arguments passed to the underlying subset method}
}
\description{
Preserves tuber metadata attributes when subsetting
}
