% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuber.R
\docType{package}
\name{tuber-package}
\alias{tuber-package}
\title{tuber: Client for the YouTube API}
\description{
Get comments posted on YouTube videos, information on how many times a video has been liked, search for videos with particular content, and much more. You can also scrape captions from a few videos. To learn more about the YouTube API, see \url{https://developers.google.com/youtube/v3/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gojiplus.github.io/tuber/}
  \item \url{https://github.com/gojiplus/tuber}
  \item Report bugs at \url{https://github.com/gojiplus/tuber/issues}
}

}
\author{
\strong{Maintainer}: Gaurav Sood \email{gsood07@gmail.com}

Authors:
\itemize{
  \item Gaurav Sood \email{gsood07@gmail.com}
}

Other contributors:
\itemize{
  \item Kate Lyons \email{k.lyons7@gmail.com} [contributor]
  \item John Muschelli \email{muschellij2@gmail.com} [contributor]
}

}
\keyword{internal}
