% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuber.R
\name{tuber_check}
\alias{tuber_check}
\title{Request Response Verification}
\usage{
tuber_check(req)
}
\arguments{
\item{req}{request}
}
\value{
in case of failure, a message
}
\description{
Request Response Verification
}
