% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuber.R
\name{tuber_DELETE}
\alias{tuber_DELETE}
\title{DELETE}
\usage{
tuber_DELETE(path, query, ...)
}
\arguments{
\item{path}{path to specific API request URL}

\item{query}{query list}

\item{\dots}{Additional arguments passed to \code{\link[httr]{DELETE}}.}
}
\value{
list
}
\description{
DELETE
}
\keyword{internal}
