% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuber.R
\name{tuber_GET}
\alias{tuber_GET}
\title{Base POST AND GET functions. Not exported.}
\usage{
tuber_GET(path, query)
}
\arguments{
\item{path}{path to specific API request URL}

\item{query}{query list}
}
\value{
list
}
\description{
GET
}

