% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuber.R
\name{tuber_GET}
\alias{tuber_GET}
\title{GET}
\usage{
tuber_GET(path, query, auth = "token", ...)
}
\arguments{
\item{path}{path to specific API request URL}

\item{query}{query list}

\item{auth}{A character vector of the authentication method, either "token" (the default) or "key"}

\item{\dots}{Additional arguments passed to \code{\link[httr]{GET}}.}
}
\value{
list
}
\description{
GET
}
