% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuber.R
\name{tuber_POST}
\alias{tuber_POST}
\title{POST}
\usage{
tuber_POST(path, query, body = "", ...)
}
\arguments{
\item{path}{path to specific API request URL}

\item{query}{query list}

\item{body}{passing image through body}

\item{\dots}{Additional arguments passed to \code{\link[httr]{GET}}.}
}
\value{
list
}
\description{
POST
}
