% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_video.R
\name{upload_video}
\alias{upload_video}
\title{Upload Video to Youtube}
\usage{
upload_video(file, snippet = list(), status = list(privacyStatus =
  "public"), query = NULL, part = "snippet,status", ...)
}
\arguments{
\item{file}{Filename of the video locally}

\item{snippet}{Additional fields for the video, including `description`
and `title`.  See
\url{https://developers.google.com/youtube/v3/docs/videos#resource} for
other fields.  Coerced to a JSON object}

\item{status}{Additional fields to be put into the \code{status} input.
options for `status` are `license` (which should hold:
`creativeCommon`, or `youtube`), `privacyStatus`, `publicStatsViewable`,
`publishAt`.}

\item{query}{Fields for `query` in `POST`}

\item{part}{The part parameter serves two purposes in this operation.
It identifies the properties that the write operation will set as
well as the properties that the API response will include.
See \url{https://developers.google.com/youtube/v3/docs/videos/insert#usage}}

\item{...}{Additional arguments to send to \code{\link{tuber_POST}} and
therefore \code{\link{POST}}}
}
\value{
A list of the response object from the \code{POST}, content,
and the URL of the uploaded
}
\description{
Upload Video to Youtube
}
\note{

}
