% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_video.R
\name{upload_video}
\alias{upload_video}
\title{Upload Video to Youtube}
\usage{
upload_video(file, snippet = list(), query = NULL,
  part = "snippet,status", ...)
}
\arguments{
\item{file}{Filename of the video locally}

\item{snippet}{Additional fields for the video, including `description`
and `title`.  See
\url{https://developers.google.com/youtube/v3/docs/videos#resource} for
other fields.  Coerced to a JSON object}

\item{query}{Fields for `query` in `POST`}

\item{part}{The part parameter serves two purposes in this operation.
It identifies the properties that the write operation will set as
well as the properties that the API response will include.
See \url{https://developers.google.com/youtube/v3/docs/videos/insert#usage}}

\item{...}{Additional arguments to send to \code{\link{tuber_POST}} and
therefore \code{\link{POST}}}
}
\value{
A response object from the \code{POST}, where the
content can be extracted using \code{httr::content}
}
\description{
Upload Video to Youtube
}
