\name{warn_deprecated}
\alias{warn_deprecated}
\title{Warn about deprecated functionality with migration guidance}
\description{Warn about deprecated functionality with migration guidance}
\usage{
warn_deprecated(old_function, new_function, version)
}
\arguments{
  \item{old_function}{Name of deprecated function}
  \item{new_function}{Name of replacement function}
  \item{version}{Version when deprecation will become an error}
}
