% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuber.R
\name{yt_key}
\alias{yt_key}
\alias{yt_get_key}
\alias{yt_set_key}
\title{Manage YouTube API key}
\usage{
yt_get_key(decrypt = FALSE)
yt_set_key(key, type)
}
\arguments{
\item{decrypt}{A boolean vector specifying whether to decrypt the supplied key with `httr2::secret_decrypt()`. Defaults to `FALSE`. If `TRUE`, requires the environment variable `TUBER_KEY` to be set in `.Renviron`.}

\item{key}{A character vector specifying a YouTube API key.}

\item{type}{A character vector specifying the type of API key to set. One of 'api' (the default, stored in `YOUTUBE_KEY`) or 'package'. Package keys are stored in `TUBER_KEY` and are used to decrypt API keys, for use in testing and continuous integration platforms.}
}
\value{
`yt_get_key()` returns a character vector with the YouTube API key stored in `.Renviron`. If this value is not stored in `.Renviron`, the functions return `NULL`.

When the `type` argument is set to 'api', `yt_set_key()` assigns a YouTube API key to `YOUTUBE_KEY` in `.Renviron` and invisibly returns `NULL`. When the `type` argument is set to 'package', `yt_set_key()` assigns a package key to `TUBER_KEY` in `.Renviron` and invisibly returns `NULL`.
}
\description{
These functions manage your YouTube API key and package key in \code{.Renviron}.
}
