% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_search.R
\name{yt_search}
\alias{yt_search}
\title{Search YouTube}
\usage{
yt_search(term = NULL)
}
\arguments{
\item{term}{search term; required; no default}
}
\value{
data.frame with 13 elements - kind, etag, id.kind, id.videoId, snippet.publishedAt, snippet.channelId, snippet.title, snippet.description, 
snippet.thumbnails.default.url, snippet.thumbnails.medium.url, snippet.thumbnails.high.url, snippet.channelTitle, snippet.liveBroadcastContent
}
\description{
Search YouTube
}
\examples{
 \dontrun{yt_search(term="Barack Obama")}
}
\references{
\url{https://console.developers.google.com/project}
}

