% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quota_management.R
\name{yt_set_quota_limit}
\alias{yt_set_quota_limit}
\title{Set Quota Limit}
\usage{
yt_set_quota_limit(limit)
}
\arguments{
\item{limit}{Integer. Daily quota limit in units}
}
\description{
Set the daily quota limit (default is 10,000 units)
}
\examples{
\dontrun{
# If you have a higher quota limit
yt_set_quota_limit(50000)
}
}
