% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuber.R
\name{yt_token}
\alias{yt_token}
\alias{yt_authorized}
\alias{yt_check_token}
\title{Check if authentication token is in options}
\usage{
yt_token()

yt_authorized()

yt_check_token()
}
\value{
A Token2.0 class
}
\description{
Check if authentication token is in options
}
