% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_topic_search.R
\name{yt_topic_search}
\alias{yt_topic_search}
\title{Search YouTube by Topic
It uses the Freebase list of topics}
\usage{
yt_topic_search(topic = NULL, ...)
}
\arguments{
\item{topic}{topic being searched for; required; no default}

\item{\dots}{Additional arguments passed to \code{\link{tuber_GET}}.}
}
\value{
a list
}
\description{
Search YouTube by Topic
It uses the Freebase list of topics
}
\examples{
 \dontrun{

# Set API token via yt_oauth() first

yt_topic_search(topic="Barack Obama")
}
}
